First pass at making delete_dataset work with published datasets

+++ src/lambda/delete_dataset/app.py
@@ -4,9 +4,13 @@ import boto3
 from botocore.client import ClientError
 
 from pydantic import BaseModel, Extra, Field, ValidationError
+from sqlalchemy import select
+from sqlalchemy.orm import Session
 from auth import check_auth
+from engine import get_engine
 from format import format_response
-from register import Dataset
+from models import PublishedDataset
+from register import Dataset, DatasetReleaseStatus
 
 DYNAMODB = boto3.resource("dynamodb")
 METADATA_TABLE = DYNAMODB.Table(os.environ["METADATA_TABLE_NAME"])
@@ -23,6 +27,23 @@ class DeleteDatasetBody(BaseModel):
 
     class Config:
         extra = Extra.forbid
+
+
+def delete_published_dataset(dataset: Dataset):
+    engine = get_engine()
+
+    with Session(engine) as session:
+        published_dataset = session.scalar(
+            select(PublishedDataset).where(
+                PublishedDataset.dataset_id == dataset.dataset_id
+            )
+        )
+
+        if not published_dataset:
+            raise ValueError("Dataset not found in database")
+
+        session.delete(published_dataset)
+        session.commit()
 
 
 def lambda_handler(event, _):
@@ -42,7 +63,21 @@ def lambda_handler(event, _):
     ):
         return format_response(403, "Not Authorized")
 
+    if validated.dataset.release_status == DatasetReleaseStatus.PUBLISHING:
+        return format_response(
+            403, "Cannot delete dataset while it is being published."
+        )
+
     try:
+        if validated.dataset.release_status == DatasetReleaseStatus.PUBLISHED:
+            delete_published_dataset(validated.dataset)
+
+    except ValueError as e:
+        print(e)
+        return format_response(403, "Error unpublishing dataset")
+
+    try:
+
         METADATA_TABLE.delete_item(
             Key={"pk": validated.dataset.project_id, "sk": validated.dataset.dataset_id}
         )
